Master merged

+++ CSH/library/api/fermat-csh-api/src/main/java/com/bitdubai/fermat_csh_api/layer/csh_wallet/cash_money/interfaces/CashMoneyManager.java
@@ -20,7 +20,7 @@ public interface CashMoneyManager {
                 ,final String status
                 ,final String balanceType
                 ,final String transactionType
-                ,final float amount
+                ,final double amount
                 ,final String cashCurrencyType
                 ,final String cashReference
                 ,final long runningBookBalance
